sweet alert delete now working

+++ src/components/resultsPage/allResults.js
@@ -2,10 +2,15 @@ import React, { Component } from 'react'
 
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 
 import axios from 'axios';
 
+
 class allResults extends Component {
+
+
 
     componentDidMount() {
         this.handleRefresh()
@@ -24,22 +29,43 @@ class allResults extends Component {
     }
     handleDelete = (id) => {
         console.log('in handle delete', id);
-        axios.delete(`/main/${id}`)
-            .then((response) => {
-                this.handleRefresh()
-            })
-            .catch((error) => {
-                alert('Error on delete');
-                console.log('Error on DELTE', error);
-            })
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.value) {
+                axios.delete(`/main/${id}`)
+                    .then((response) => {
+                        this.handleRefresh()
+                    })
+                    .catch((error) => {
+                        alert('Error on delete');
+                        console.log('Error on DELTE', error);
+                    })
+                Swal.fire(
+                    'Smart Mark Deleted!',
+
+                    'success'
+                )
+            }
+        })
+
     }
+
+
+
 
     displayItems = (list) => {
 
         if (list) {
             return (
                 <div>
-                    <header> RESULTS</header>
+                    <h1>  ALL MARKS</h1>
                     {list.map(item =>
                         <table>
                             <tr>
@@ -47,6 +73,9 @@ class allResults extends Component {
                                 <th>Error-Code</th>
                                 <th>Site-Name</th>
                                 <th>link</th>
+                                <th>delete result </th>
+                                <th>edit result</th>
+
                             </tr>
                             <tr>
                                 <td>{item.topic}</td>
@@ -54,19 +83,22 @@ class allResults extends Component {
                                 <td>{item.site}</td>
 
                                 <td><a href={item.url}>{item.url}</a></td>
-                                <button onClick={() => this.handleEdit(item)}>Edit</button>
-                                <button onClick={() => this.handleDelete(item.id)}>Delete</button>
+                                <td> <button onClick={() => this.handleEdit(item)}>Edit</button></td>
+                                <td>   <button onClick={() => this.handleDelete(item.id)}>Delete</button></td>
                             </tr>
                         </table>
-                    )}
-                </div>
+                    )
+                    }
+                </div >
             )
 
 
         }
     }
     render() {
+
         return (
+
             <div>
                 {this.displayItems(this.props.getItem)}
             </div>
